Add files via upload
In some of the non  metropolitan cities like Visakhapatnam there is no proper online bike rental service. Today in this busy schedule, people do not get time to go to shop and schedule their appointment. Bike  rentals is a web-based bike rental management application with bike booking functionality. It connects clients and bike rental shop in an online community allowing users to browse bikes and brands. Bike  rentals  is on-demand, easy to use, and effective for users and bike rentals shops.